Fix failing log test handler tests
The #42751 removed airflow_version from k8s log handler and
apparently some tests in tests_utils still use kubernetes provider
and k8s test handler. Also it turned out some other tests
used Celery Executor as well.

Fixed the tests and move them to K8S / Celery provider respectively.
